+++ app/entries/entries.module.ts
@@ -3,7 +3,7 @@ import { Controller } from '@hals/common';
 import { EntriesService } from './entries.service';
 import { MongoEntriesRepository } from './mongo-entries-repository.service';
 import { EntriesValidator } from './entries.validator';
-import { MongoUsersRepository } from '../users/mongo-users-repository';
+import { MongoUsersRepository } from '../users/mongo-users-repository.service';
 import { MongoJournalsRepository } from '../journals/mongo-journals-repository.service';
 
 const entriesService : EntriesService = EntriesService(
