+++ app/journals/journals.module.ts
@@ -1,6 +1,6 @@
 import { JournalsService } from './journals.service';
 import { MongoJournalsRepository } from './mongo-journals-repository.service';
-import { MongoUsersRepository } from '../users/mongo-users-repository';
+import { MongoUsersRepository } from '../users/mongo-users-repository.service';
 import { JournalsValidator } from './journals.validator';
 import { Controller } from '@hals/common';
 import { JournalsController } from './journals.controller';
